las acciones del DeveloperView ya traen las tareas del usuario y las que necesitan ayuda, desde la base de datos
