+++ client/src/redux/DeveloperView/constants.js
@@ -1,5 +1,6 @@
 //Constantes
 export const CONSTANTS = {
     GET_USER_TASKS: "GET_USER_TASKS",
-    GET_HELP_TASKS: "GET_HELP_TASKS"
+    GET_HELP_TASKS: "GET_HELP_TASKS",
+    SET_USER: "SET_USER"
   }
