+++ client/src/redux/DeveloperView/reducer.js
@@ -2,7 +2,8 @@ import { CONSTANTS } from './constants.js';
 
 const initialState = {
   userTasks: [],
-  helpTasks: []
+  helpTasks: [],
+  loggedUser: {},
 };
 
 export default function reducer(state = initialState, action) {
@@ -10,7 +11,7 @@ export default function reducer(state = initialState, action) {
     case CONSTANTS.GET_USER_TASKS:
       return {
         ...state,
-        userTasks: action.payload
+        userTasks: action.payload.filter((task) => task.asignedTo !== state.loggedUser._id),
       }
     case CONSTANTS.GET_HELP_TASKS:
       return {
