+++ client/src/redux/rootReducer.js
@@ -1,12 +1,11 @@
 import { combineReducers } from "redux";
 
-import developerViewReducer from "./DeveloperView/DeveloperViewReducer";
+import developerViewReducer from "./DeveloperView/reducer.js";
 import managerViewReducer from "./ManagerView/reducer.js";
 import homeReducer from "./Home/reducer";
 
 const rootReducer = combineReducers({
   developerView: developerViewReducer,
-  // Aca van agregando los siguientes reducers
   managerView: managerViewReducer,
   home: homeReducer,
 });
